+++ app/initializers/ember-validations.js
@@ -1,0 +1,1 @@
+export { default, initialize } from 'ember-flexberry/initializers/ember-validations';