+++ tests/dummy/app/routes/components-acceptance-tests/flexberry-lookup/settings-example-custom-window.js
@@ -1,0 +1,34 @@
+import Ember from 'ember';
+import EditFormRoute from 'ember-flexberry/routes/edit-form';
+
+export default EditFormRoute.extend({
+  /**
+    Name of model projection to be used as record's properties limitation.
+
+    @property modelProjection
+    @type String
+    @default 'SettingLookupExampleView'
+   */
+  modelProjection: 'SuggestionE',
+
+  /**
+    Name of model to be used as form's record type.
+
+    @property modelName
+    @type String
+    @default 'ember-flexberry-dummy-suggestion'
+   */
+  modelName: 'ember-flexberry-dummy-suggestion',
+
+  /**
+    Returns model related to current route.
+
+    @method model
+   */
+  model() {
+    let store = Ember.get(this, 'store');
+    return store.createRecord('ember-flexberry-dummy-suggestion', {
+      name: 'TestSuggestionName'
+    })
+  }
+});